requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
six==1.16.0
spotipy==2.23.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
